Fix Path to Aragen to recipe specific sysroot
ARA_GENERATOR_DIR is defined in some recipes from meta-st layer.
It is structured as follows: ${STAGING_DIR}/${BUILD_SYS}/usr/share/ara-gen/
where "STAGING_DIR" is the path to "build_dir/tmp" and "BUILD_SYS" is
expanded to the host architecture and OS: "x86_64-linux".
This doesn't work in rocko distribution, where every file used in by a
built application is deployed in a recipe specific sysroot.
Therefore, we change the "BUILD_SYS" variable to point to the sysroot of
each recipe that uses it.

Issue: ALB-3298
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>