sumo: Set LAYERSERIES_COMPAT to match sumo version
In yocto sumo release, layers are expected to set
LAYERSERIES_COMPAT variable in order to match the version(s)
of OpenEmbedded-Core they are compatible with.
Therefore, we set this variable to match the version sumo,
in order to avoid any warnings that can be generated at
build time, due to upgrading poky to sumo.

Issue: ALB-3511
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>